feat: memory network reliability
