+++ crates/hotshot/src/traits.rs
@@ -14,7 +14,7 @@ pub mod implementations {
     pub use super::{
         networking::{
             libp2p_network::{Libp2pCommChannel, Libp2pNetwork, PeerInfoVec},
-            memory_network::{DummyReliability, MasterMap, MemoryCommChannel, MemoryNetwork},
+            memory_network::{MasterMap, MemoryCommChannel, MemoryNetwork},
             web_server_libp2p_fallback::{CombinedNetworks, WebServerWithFallbackCommChannel},
             web_server_network::{WebCommChannel, WebServerNetwork},
         },
