+++ crates/types/src/traits/network.rs
@@ -2,9 +2,10 @@
 //!
 //! Contains types and traits used by `HotShot` to abstract over network access
 
+use async_compatibility_layer::art::async_sleep;
 #[cfg(async_executor_impl = "async-std")]
 use async_std::future::TimeoutError;
-use hotshot_task::BoxSyncFuture;
+use hotshot_task::{boxed_sync, BoxSyncFuture};
 use libp2p_networking::network::NetworkNodeHandleError;
 #[cfg(async_executor_impl = "tokio")]
 use tokio::time::error::Elapsed as TimeoutError;
@@ -14,6 +15,10 @@ use super::{election::Membership, node_implementation::NodeType, signature_key::
 use crate::{data::ViewNumber, message::MessagePurpose};
 use async_compatibility_layer::channel::UnboundedSendError;
 use async_trait::async_trait;
+use rand::{
+    distributions::{Bernoulli, Uniform},
+    prelude::Distribution,
+};
 use serde::{Deserialize, Serialize};
 use snafu::Snafu;
 use std::{collections::BTreeSet, fmt::Debug, sync::Arc, time::Duration};
@@ -366,6 +371,7 @@ pub enum NetworkChange<P: SignatureKey> {
 }
 
 /// interface describing how reliable the network is
+#[async_trait]
 pub trait NetworkReliability: Debug + Sync + std::marker::Send {
     /// Sample from bernoulli distribution to decide whether
     /// or not to keep a packet
@@ -373,8 +379,225 @@ pub trait NetworkReliability: Debug + Sync + std::marker::Send {
     ///
     /// Panics if `self.keep_numerator > self.keep_denominator`
     ///
-    fn sample_keep(&self) -> bool;
+    fn sample_keep(&self) -> bool {
+        true
+    }
+
     /// sample from uniform distribution to decide whether
     /// or not to keep a packet
-    fn sample_delay(&self) -> Duration;
+    fn sample_delay(&self) -> Duration {
+        std::time::Duration::ZERO
+    }
+
+    /// scramble the packet
+    fn scramble(&self, msg: Vec<u8>) -> Vec<u8> {
+        msg
+    }
+
+    /// number of times to repeat the packet
+    fn sample_repeat(&self) -> usize {
+        1
+    }
+
+    /// given a message and a way to send the message,
+    /// decide whether or not to send the message
+    /// how long to delay the message
+    /// whether or not to send duplicates
+    /// and whether or not to include noise with the message
+    /// then send the message
+    fn chaos_send_msg(
+        &self,
+        msg: Vec<u8>,
+        send_fn: Arc<dyn Send + Sync + 'static + Fn(Vec<u8>) -> BoxSyncFuture<'static, ()>>,
+    ) -> BoxSyncFuture<'static, ()> {
+        let sample_keep = self.sample_keep();
+        let delay = self.sample_delay();
+        let repeats = self.sample_repeat();
+        let mut msgs = Vec::new();
+        for _idx in 0..repeats {
+            let scrambled = self.scramble(msg.clone());
+            msgs.push(scrambled);
+        }
+        let closure = async move {
+            if sample_keep {
+                async_sleep(delay).await;
+                for msg in msgs {
+                    send_fn(msg).await;
+                }
+            }
+        };
+        boxed_sync(closure)
+    }
+}
+
+/// ideal network
+#[derive(Clone, Copy, Debug, Default)]
+pub struct PerfectNetwork {}
+
+impl NetworkReliability for PerfectNetwork {}
+
+/// A synchronous network. Packets may be delayed, but are guaranteed
+/// to arrive within `timeout` ns
+#[derive(Clone, Copy, Debug, Default)]
+pub struct SynchronousNetwork {
+    /// Max delay of packet before arrival
+    timeout_ms: u64,
+    /// Lowest value in milliseconds that a packet may be delayed
+    delay_low_ms: u64,
+}
+
+impl NetworkReliability for SynchronousNetwork {
+    /// never drop a packet
+    fn sample_keep(&self) -> bool {
+        true
+    }
+    fn sample_delay(&self) -> Duration {
+        Duration::from_millis(
+            Uniform::new_inclusive(self.delay_low_ms, self.timeout_ms)
+                .sample(&mut rand::thread_rng()),
+        )
+    }
+}
+
+/// An asynchronous network. Packets may be dropped entirely
+/// or delayed for arbitrarily long periods
+/// probability that packet is kept = `keep_numerator` / `keep_denominator`
+/// packet delay is obtained by sampling from a uniform distribution
+/// between `delay_low_ms` and `delay_high_ms`, inclusive
+#[derive(Debug, Clone, Copy)]
+pub struct AsynchronousNetwork {
+    /// numerator for probability of keeping packets
+    keep_numerator: u32,
+    /// denominator for probability of keeping packets
+    keep_denominator: u32,
+    /// lowest value in milliseconds that a packet may be delayed
+    delay_low_ms: u64,
+    /// highest value in milliseconds that a packet may be delayed
+    delay_high_ms: u64,
+}
+
+impl NetworkReliability for AsynchronousNetwork {
+    fn sample_keep(&self) -> bool {
+        Bernoulli::from_ratio(self.keep_numerator, self.keep_denominator)
+            .unwrap()
+            .sample(&mut rand::thread_rng())
+    }
+    fn sample_delay(&self) -> Duration {
+        Duration::from_millis(
+            Uniform::new_inclusive(self.delay_low_ms, self.delay_high_ms)
+                .sample(&mut rand::thread_rng()),
+        )
+    }
+}
+
+/// An partially synchronous network. Behaves asynchronously
+/// until some arbitrary time bound, GST,
+/// then synchronously after GST
+#[allow(clippy::similar_names)]
+#[derive(Debug, Clone, Copy)]
+pub struct PartiallySynchronousNetwork {
+    /// asynchronous portion of network
+    asynchronous: AsynchronousNetwork,
+    /// synchronous portion of network
+    synchronous: SynchronousNetwork,
+    /// time when GST occurs
+    gst: std::time::Duration,
+    /// when the network was started
+    start: std::time::Instant,
+}
+
+impl NetworkReliability for PartiallySynchronousNetwork {
+    /// never drop a packet
+    fn sample_keep(&self) -> bool {
+        true
+    }
+    fn sample_delay(&self) -> Duration {
+        // act asyncronous before gst
+        if self.start.elapsed() < self.gst {
+            if self.asynchronous.sample_keep() {
+                self.asynchronous.sample_delay()
+            } else {
+                // assume packet was "dropped" and will arrive after gst
+                self.synchronous.sample_delay() + self.gst
+            }
+        } else {
+            // act syncronous after gst
+            self.synchronous.sample_delay()
+        }
+    }
+}
+
+impl Default for AsynchronousNetwork {
+    // disable all chance of failure
+    fn default() -> Self {
+        AsynchronousNetwork {
+            keep_numerator: 1,
+            keep_denominator: 1,
+            delay_low_ms: 0,
+            delay_high_ms: 0,
+        }
+    }
+}
+
+impl Default for PartiallySynchronousNetwork {
+    fn default() -> Self {
+        PartiallySynchronousNetwork {
+            synchronous: SynchronousNetwork::default(),
+            asynchronous: AsynchronousNetwork::default(),
+            gst: std::time::Duration::new(0, 0),
+            start: std::time::Instant::now(),
+        }
+    }
+}
+
+impl SynchronousNetwork {
+    /// create new `SynchronousNetwork`
+    #[must_use]
+    pub fn new(timeout: u64, delay_low_ms: u64) -> Self {
+        SynchronousNetwork {
+            timeout_ms: timeout,
+            delay_low_ms,
+        }
+    }
+}
+
+impl AsynchronousNetwork {
+    /// create new `AsynchronousNetwork`
+    #[must_use]
+    pub fn new(
+        keep_numerator: u32,
+        keep_denominator: u32,
+        delay_low_ms: u64,
+        delay_high_ms: u64,
+    ) -> Self {
+        AsynchronousNetwork {
+            keep_numerator,
+            keep_denominator,
+            delay_low_ms,
+            delay_high_ms,
+        }
+    }
+}
+
+impl PartiallySynchronousNetwork {
+    /// create new `PartiallySynchronousNetwork`
+    #[allow(clippy::similar_names)]
+    #[must_use]
+    pub fn new(
+        asynchronous: AsynchronousNetwork,
+        synchronous: SynchronousNetwork,
+        gst: std::time::Duration,
+    ) -> Self {
+        PartiallySynchronousNetwork {
+            asynchronous,
+            synchronous,
+            gst,
+            start: std::time::Instant::now(),
+        }
+    }
+}
+
+/// A chaotic network using all the networking calls
+pub struct ChaosNetwork {
+    // TODO
 }
